Add Windows compatibility stuff

+++ client/engine/Vector2d.c
@@ -1,3 +1,4 @@
+#include "../global.h"
 #include "Vector2d.h"
 
 #include <math.h>
@@ -32,7 +33,7 @@ float Vector2d_magnitude_squared(Vector2d v) {
 }
 
 float Vector2d_magnitude(Vector2d v) {
-    return sqrt(Vector2d_magnitude_squared(v));
+    return (float) sqrt(Vector2d_magnitude_squared(v));
 }
 
 float Vector2d_angle(Vector2d v) {
